refactor pkg
